+++ src/main/java/com/skyflow/entities/LogLevel.java
@@ -5,4 +5,5 @@ public enum LogLevel {
     INFO,
     WARN,
     ERROR,
+    OFF
 }